# api/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.ErrorBody:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
  apidocs.WatermarkAssetData:
    properties:
      data:
        $ref: '#/definitions/domain.WatermarkAsset'
    type: object
  apidocs.WatermarkAssetList:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.WatermarkAsset'
        type: array
      dir:
        type: string
      total:
        type: integer
    type: object
...
  domain.WatermarkAsset:
    properties:
      content_type:
        description: |-
          ContentType is the MIME type sniffed at upload time
          (e.g. "image/png", "image/jpeg"). Used for Content-Type on /raw GET.
        type: string
      file_name:
        description: |-
          FileName is the original filename at upload time, preserved for
          audit / download responses. Includes extension (e.g. "logo.png").
        type: string
      id:
        description: ID is the immutable filesystem-safe identifier (also the on-disk
          basename).
        type: string
      name:
        description: |-
          Name is the human display label shown in UIs. Defaults to the upload's
          original filename when the operator doesn't override.
        type: string
      size_bytes:
        description: SizeBytes is the on-disk size of the image. Sidecar JSON is excluded.
        type: integer
      uploaded_at:
        description: UploadedAt is the wall-clock UTC time the asset was first stored.
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for live stream configuration, ingest/publish lifecycle, DVR
    recordings, and event hooks.
  title: Open Streamer API
  version: "1.0"
paths:
...
  /watermarks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List watermark assets
      tags:
      - watermarks
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image file (PNG / JPG / GIF, ≤ 8 MiB)
        in: formData
        name: file
        required: true
        type: file
      - description: Display name (defaults to filename)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Upload a watermark image
      tags:
      - watermarks
  /watermarks/{id}:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete a watermark asset
      tags:
      - watermarks
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get watermark metadata
      tags:
      - watermarks
  /watermarks/{id}/raw:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Download watermark image
      tags:
      - watermarks
swagger: "2.0"